Revert "refactor(sidekick): move `rustBumpVersions()` "
This reverts commit 0f16117e05a3047fca596759a25f1d595eb4cb68.
