--- internal/sidekick/rust_bump_versions_test.go
@@ -1,35 +1,0 @@
-// Copyright 2025 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package sidekick
-
-import (
-	"testing"
-
-	"github.com/googleapis/librarian/internal/sidekick/internal/config"
-)
-
-func TestRustBumpVersions(t *testing.T) {
-	config := &config.Config{
-		Release: &config.Release{
-			Preinstalled: map[string]string{
-				"git": "git-not-found",
-			},
-		},
-	}
-	cmdLine := &CommandLine{}
-	if err := rustBumpVersions(config, cmdLine); err == nil {
-		t.Errorf("expected an error with invalid git command")
-	}
-}
