+++ internal/sidekick/sidekick_test.go
@@ -49,3 +49,7 @@ func requireProtoc(t *testing.T) {
 func requireCargo(t *testing.T) {
 	requireCommand(t, "cargo")
 }
+
+func requireGit(t *testing.T) {
+	requireCommand(t, "git")
+}
